requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
boolean.py==4.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
cryptography==39.0.0
databases==0.7.0
dnspython==2.2.1
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.88.0
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
jsonschema==4.17.3
jsonschema-spec==0.1.2
kafka-python==2.0.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.4
multipart==0.2.4
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==22.0
passlib==1.7.4
pathable==0.4.3
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
pymongo==4.3.3
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
swagger-ui-bundle==0.0.9
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
var-dump==1.2
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.2
zipp==3.11.0
